+++ v2/tools/testdata/gen_fromdb_sqlite/idl.go
@@ -29,7 +29,7 @@ type UserItem struct {
 type CreateUserRequest struct {
 	Username string `json:"username"`
 	Email string `json:"email"`
-	CreatedAt *time.Time `json:"created_at"`
+	CreatedAt *time.Time `json:"created_at,omitempty"`
 }
 
 // CreateUserResponse 创建响应